pyproject.toml
[project]
name = "pydantic-ai-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "langchain[anthropic]>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.10",
    "llm-sandbox>=0.3.12",
    "pandas>=2.3.0",
    "pydantic-ai>=0.3.1",
    "pydantic[email,timezone]>=2.11.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scipy>=1.15.3",
    "vllm>=0.9.1",
]

[tool.uv.sources]
pydantic = { git = "https://github.com/pydantic/pydantic", rev = "main" }
